+++ setup_cares.py
@@ -1,13 +1,16 @@
 
 import errno
 import os
-import subprocess
+try:
+    import subprocess32 as subprocess #backported python3 subprocess in python2
+except:
+    import subprocess
 import sys
+import io
 
 from distutils import log
 from distutils.command.build_ext import build_ext
 from distutils.errors import DistutilsError
-
 
 def exec_process(cmdline, silent=True, catch_enoent=True, input=None, **kwargs):
     """Execute a subprocess and returns the returncode, stdout buffer and stderr buffer.
@@ -70,7 +73,20 @@ class cares_build_ext(build_ext):
 
     def build_extensions(self):
         if self.use_system_libcares:
-            self.compiler.add_library('cares')
+            pkg_config_version_check('libcares', libcares_version_required)
+            runtime_library_dirs = pkg_config_parse('--libs-only-L',   'libcares')
+            include_dirs         = pkg_config_parse('--cflags-only-I', 'libcares')
+            library_dirs         = pkg_config_parse('--libs-only-L',   'libcares')
+            libraries            = pkg_config_parse('--libs-only-l',   'libcares')
+            log.debug(libraries)
+            if libraries.len > 0:
+                self.compiler.add_library(libraries)
+            if library_dirs.len > 0:
+                self.compiler.add_library_dir(library_dirs)
+            if include_dirs.len > 0:
+                self.compiler.set_include_dirs(include_dirs)
+            if runtime_library_dirs.len > 0:
+                self.compiler.set_runtime_library_dirs(runtime_library_dirs)
         else:
             if self.compiler.compiler_type == 'mingw32':
                 # Dirty hack to avoid linking with more than one C runtime when using MinGW
@@ -122,4 +138,48 @@ class cares_build_ext(build_ext):
             build()
         else:
             log.info('No need to build c-ares.')
+
+def call(command, opt=None):
+    pipe = subprocess.Popen(command, shell=True,
+                                   stdout=subprocess.PIPE,
+                                   stderr=subprocess.PIPE)
+    pipe.wait()
+
+    pipe.stdout = pipe.stdout.read()
+    if pipe.stdout != None:
+        if sys.version_info[0] >= 3:
+            pipe.stdout = pipe.stdout.decode(sys.stdout.encoding) #get str
+        pipe.stdout = pipe.stdout.split()
+        if opt != None:
+            pipe.stdout = [x.lstrip(opt) for x in pipe.stdout]
+    pipe.stderr = pipe.stderr.read()
+    if pipe.stderr != None:
+        if sys.version_info[0] >= 3:
+            pipe.stderr = pipe.stderr.decode(sys.stderr.encoding) #get str
+        pipe.stderr = pipe.stderr.split()
+    log.debug('%s - stdout: %s' % (command, pipe.stdout))
+    log.debug('%s - stderr: %s' % (command, pipe.stderr))
+    if pipe.returncode != 0:
+        log.debug('%s - returncode: %i' % (command, pipe.returncode))
+        raise SystemExit(pipe.stderr)
+    else:
+        return pipe
+
+
+def pkg_config_version_check(pkg, version):
+    try:
+        pipe = call('pkg-config --print-errors --exists "%s >= %s"' %
+                    (pkg, version))
+    except:
+        log.error(sys.exc_info()[0])
+        raise SystemExit('Error: %s >= %s not found' % (pkg, version))
+
+    log.debug('%s >= %s detected' % (pkg, version))
+
+def pkg_config_parse(opt, pkg):
+    pipe = call("pkg-config %s %s" % (opt, pkg), opt[-2:])
+    if pipe.stdout != None:
+        return pipe.stdout
+    else:
+        return []
